+++ src/components/RecipeNewForm/SliderComponent.tsx
@@ -16,40 +16,43 @@ export default function SliderComponent() {
     };
 
     return (
-        <div className={classes.ItemDescription}>
-            <div className={classes.sliderData}>
-                <span className={classes.minValue}> {'<'}10</span>
-                <MuiInput
-                    value={value}
-                    size="small"
-                    id="value-input"
-                    disableUnderline
-                    className={classes.CurrentValue}
-                    onChange={handleInputChange}
-                    sx={{ color: '#1FCC79', fontSize: '15px', fontWeight: 'bold' }}
-                />
+        <div className={classes.Item}>
+            <span className={classes.ItemTitle}>Время готовки (в минутах)</span>
+            <div className={classes.ItemDescription}>
+                <div className={classes.sliderData}>
+                    <span className={classes.minValue}> {'<'}10</span>
+                    <MuiInput
+                        value={value}
+                        size="small"
+                        id="value-input"
+                        disableUnderline
+                        className={classes.CurrentValue}
+                        onChange={handleInputChange}
+                        sx={{ color: '#1FCC79', fontSize: '15px', fontWeight: 'bold' }}
+                    />
 
-                <span className={classes.maxValue}> {'>'} 60 </span>
+                    <span className={classes.maxValue}> {'>'} 60 </span>
+                </div>
+                <Slider
+                    max={60}
+                    defaultValue={30}
+                    onChange={handleSliderChange}
+                    sx={{
+                        color: '#1FCC79',
+                        '.MuiSlider-rail': {
+                            color: '#F4F5F7',
+                            height: '8px',
+                        },
+                        '.MuiSlider-thumb': {
+                            width: '40px',
+                            borderRadius: '32px',
+                        },
+                        '.MuiSlider-track': {
+                            height: '8px',
+                        },
+                    }}
+                />
             </div>
-            <Slider
-                max={60}
-                defaultValue={30}
-                onChange={handleSliderChange}
-                sx={{
-                    color: '#1FCC79',
-                    '.MuiSlider-rail': {
-                        color: '#F4F5F7',
-                        height: '8px',
-                    },
-                    '.MuiSlider-thumb': {
-                        width: '40px',
-                        borderRadius: '32px',
-                    },
-                    '.MuiSlider-track': {
-                        height: '8px',
-                    },
-                }}
-            />
         </div>
     );
 }
